requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
tenacity==8.2.3
pydantic==2.5.3
